// src/model/notification/notification-type.ts
export enum NOTIFICATION_TYPE {
    DIRECT_CALL = 'DIRECT_CALL',
    ANONYMOUS_CALL = 'ANONYMOUS_CALL',
    ANONYMOUS_MESSAGE = 'ANONYMOUS_MESSAGE',
    DIRECT_MESSAGE = 'DIRECT_MESSAGE',
    FRIEND_REQUEST = 'FRIEND_REQUEST',
    FRIEND_REQUEST_ACCEPTED = 'FRIEND_REQUEST_ACCEPTED',
    REPORT_HANDLED = 'REPORT_HANDLED',
}